% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{a_b_bagplot}
\alias{a_b_bagplot}
\title{Creates bagplots of equitable intercepts versus 1-slope}
\usage{
a_b_bagplot(community.f = NULL, Td, refindex = "Row_Ave",
  xlim = NULL, ylim = NULL, main = " ", rownum = NULL,
  facname = NULL, fall = FALSE)
}
\arguments{
\item{community.f}{nonnumeric character traits of the individuals in x: default NULL}

\item{Td}{Transform info based on transformE}

\item{refindex}{intercept values based on this refences x location or individual: Default "Row_Ave"}

\item{xlim}{vector of c(lowest range,highest range) for 1-slope values default NULL function defines them}

\item{ylim}{vector of c(lowest range,highest range) for intercept values default NULL function defines them}

\item{main}{text heading for plot}

\item{rownum}{vector of points on t to use as zeroes default NULL (about 10 t values chosen)}

\item{facname}{name of factor characteristic}

\item{fall}{use all values of t Default=FALSE}
}
\value{
index of best intersection of plotted sequences
}
\description{
Bagplots are made by varying the zero of ther data set to various values the variable t(dependent on options).
If community.f has some nonnumeric trait then bagplots are also created for each category of the trait.
Plots are made based on the reference column chosen.
}
\examples{

d<-eg8(rmult=4,cmult=4,mf=10,mg= 0,mu=10,sdf=0.75,sdg=1,sdu=0.2)
Td<-transformE(d=d)
aa<-a_b_bagplot(Td=Td,xlim=c(-0.3,0.3),ylim=c(-4,4))
aa<-a_b_bagplot(Td=Td,xlim=c(-0.3,0.3),ylim=c(-5,5),rownum=seq(1,ncol(Td$smat),by=2))
aa<-a_b_bagplot(Td=Td,xlim=c(-0.2,0.2),ylim=c(-5,5),fall=TRUE)

}
