% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{eg4}
\alias{eg4}
\title{Example 4 of equitable system
f(x) product of sqrt of column index and very large scale wave:
u(x) linear in column index
g(x) sum of two  waves}
\usage{
eg4(rmult = 1, cmult = 1)
}
\arguments{
\item{rmult}{resolution of sampling of t variable  rmult*15 is the number of points sampled}

\item{cmult}{resolution of sampling of x variable  cmult*10 is the number of points sampled}
}
\value{
matrix of equitable data
}
\description{
Example 4 of equitable system
f(x) product of sqrt of column index and very large scale wave:
u(x) linear in column index
g(x) sum of two  waves
}
\examples{
d<-eg4(1,1)       # 15x10 data set
d<-eg4(1,10)      # 15x100 data set
d<-eg4(10,1)      # 150x10 data set
d<-eg4(15,15)    # 150x100 data set

}
