% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{eg5}
\alias{eg5}
\title{Example 5 of equitable system
f(x) product of sqrt of column index and short wavelength wave:
u(x) sum of linear in column index with short wavelength
g(x) sum of two  waves}
\usage{
eg5(rmult = 1, cmult = 1)
}
\arguments{
\item{rmult}{resolution of sampling of t variable  rmult*15 is the number of points sampled}

\item{cmult}{resolution of sampling of x variable  cmult*10 is the number of points sampled}
}
\value{
matrix of equitable data
}
\description{
Example 5 of equitable system
f(x) product of sqrt of column index and short wavelength wave:
u(x) sum of linear in column index with short wavelength
g(x) sum of two  waves
}
\examples{
d<-eg5(1,1)       # 15x10 data set
d<-eg5(1,10)      # 15x100 data set
d<-eg5(10,1)      # 150x10 data set
d<-eg5(15,15)    # 150x100 data set

}
