% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{eg8}
\alias{eg8}
\title{Example 8 of equitable system}
\usage{
eg8(rmult = 1, cmult = 1, mf = 10, mg = 0, mu = 10, sdf = 0.75,
  sdg = 1, sdu = 0.2)
}
\arguments{
\item{rmult}{resolution of sampling of t variable  rmult*15 is the number of points sampled}

\item{cmult}{resolution of sampling of x variable  cmult*10 is the number of points sampled}

\item{mf}{10 (default) mean of f(x)}

\item{mg}{0 (default) mean of g(x)}

\item{mu}{10 (default) mean of u(x)}

\item{sdf}{0.75 (default) standard deviation of f(x)}

\item{sdg}{1 (default) standard deviation of g(x)}

\item{sdu}{0.2 (default) standard deviation of u(x)}
}
\value{
matrix of equitable data
}
\description{
f(x) , g(t) amd u(x) are samopled from normal random distributions.
g is then ordered from  lowest to highest
Good for simulating equitable phenology systems
}
\details{
f(x) random normal distribution with mean mf and standard deviation sdf
u(x) random normal distribution with mean mu and standard deviation sdu
g(x) random normal distribution with mean mg and standard deviation sdg
}
\examples{
d<-eg8(rmult=4,cmult=4)       # 60x40 data set
 #transform this data set and then show bagplot intersetions
Td<-transformE(d=d)
aa<-a_b_bagplot(Td=Td,xlim=c(-0.25,0.25),ylim=c(-5,5))

d<-eg8(3,3)       # 45x30 data set
d<-eg8(1,10)      # 15x100 data set
d<-eg8(10,1)      # 150x10 data set
d<-eg8(15,15)    # 150x100 data set

}
