% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{egsimple}
\alias{egsimple}
\title{Example "simple" of equitable system
f(x) , g(t) amd u(x) are define directly as vectors to reate equitable system}
\usage{
egsimple(f = c(2, 4, 6, 8), g = c(-6, -3, 0, 1, 2), u = c(4, 3, 0,
  1))
}
\arguments{
\item{f}{vector f(x) default c(2,4,6,8)}

\item{g}{vector g(t) default c(-6,-3,0,1,2)}

\item{u}{vector u(x) of same length as default f c(4,3,0,1) when NULL it is set to 0}
}
\value{
matrix of equitable data
}
\description{
Example "simple" of equitable system
f(x) , g(t) amd u(x) are define directly as vectors to reate equitable system
}
\examples{
egsimple()       # u(x)=c(4,3,0,1)
egsimple(u=NULL)      # u(x)=0
d<-egsimple(f=c(1,2,3),g=c(-1,0,1,2,3),u=c(5,-2,6))
d<-egsimple(f=c(1,2,3),g=c(-1,0,1,2,3),u=NULL)

}
