\name{Equitable-package}
\alias{Equitable-package}
\alias{Equitable}
\docType{package}
\title{
\packageTitle{Equitable}
}
\description{
\packageDescription{Equitable}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Equitable}
\packageIndices{Equitable}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Equitable}

Maintainer: \packageMaintainer{Equitable}
}
\references{
Elphinstone , c and Henry G. submitted to Ecology , 2018
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
