% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{make_A}
\alias{make_A}
\title{Construct equitable slope matrix from f(x) : image is displayed of matrix}
\usage{
make_A(xref = 1, f)
}
\arguments{
\item{xref}{1 (default) x reference value around which the matrix is constructed}

\item{f}{a vector f(x) around which the equitbel matirx is constructed}
}
\value{
Equitable matrix
}
\description{
Given f(x) assumed to be defined relative to x=xref, an equitable matrix is constructed
}
\examples{
make_A(xref=1,f=c(1,2,3))  #simple example of equitable matrix
make_A(xref=3,f=c(1,2,3))  #reference not important
plot(eg4(3,3)[,1])
A<-make_A(xref=1,f=eg4(3,3)[,1])
#more complicated matrix based on large scale wave f(1)g(t)  in eg4
A<-make_A(xref=2,f=eg4(3,3)[20,])
 #more complicated matrix based on large scale wave f(x)g(20) in eg4
A<-make_A(f=eg5(6,6)[,1])  #more complicated matrix based on small scale wave f(1)g(t)  in eg5
plot(eg5(10,10)[20,],type="b")
 #more complicated matrix based on small scale wave f(x)g(20) in eg5
A<-make_A(xref=2,f=eg5(10,10)[20,])
 #low resolution  matrix based on small scale wave f(x)g(20) in eg5
A<-make_A(xref=2,f=eg5(2,2)[20,])
#more complicated matrix based on small scale wave f(20)g(t) in eg7
A<-make_A(xref=2,f=eg7(10,10)[,20])
#more complicated matrix based on small scale wave f(x)g(20) in eg7
A<-make_A(xref=2,f=eg7(10,10)[20,])


}
