% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{make_B}
\alias{make_B}
\title{Construct Shift equitable matrix}
\usage{
make_B(xref = 1, orig, A, uflag = FALSE, pflag = TRUE)
}
\arguments{
\item{xref}{Reference "location" to base shift matrix default set to 1 (for PCA dominant is first)}

\item{orig}{2D Data set used when uflag FALSE   profile at t=constant when uflag=TRUE  (u(x) equivalent)}

\item{A}{Slope equitable matrix}

\item{uflag}{TRUE(default) or FALSE   TRUE uses origas u(x): FALSE uses average profile (averaged over rows)}

\item{pflag}{ignore}
}
\value{
B shift matrix B
}
\description{
Finds a shift equitbale matrix from a slope equitbel matrix and a sequence profile
}
\examples{
d<-eg4(2,2); Td<-transformE(d)
B<-make_B(orig=Td$ET.x, A=Td$E.s)
B<-make_B(xref=5,orig=Td$ET.x, A=Td$E.s)
B<-make_B(xref=5,orig=Td$ET.x[1,], A=Td$E.s,uflag=TRUE)

}
