% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{nastat}
\alias{nastat}
\title{mean and std dev on NA data set}
\usage{
nastat(d)
}
\arguments{
\item{d}{data set}
}
\value{
list of mean $m and standard deviation $std
}
\description{
Finds mean and standard deviaiton of 2 dimensional data set that contains NA values
}
\examples{
nastat(d=data.frame(cbind(c(1:4),c(2,5,NA,NA) )))  #
nastat(eg4())   # stats and image of example 4
nastat(eg7())   # stats and image of example 7


}
