% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{plotsome}
\alias{plotsome}
\title{Various types of plots of Equitable transform data dependent on options chosen}
\usage{
plotsome(T, num = 5, signal = NULL, limits = NULL, xlimits = NULL,
  indiv = FALSE, columns = FALSE, images = TRUE, density = FALSE,
  versus = FALSE, xvsref = NULL, yline = 3, yma = 5,
  transpose = FALSE, of = FALSE, lf = FALSE, ef = TRUE,
  avef = FALSE, errb = FALSE, row_unit = NULL, col_unit = NULL,
  z_unit = NULL, genname = NULL, stderror = FALSE, fcontour = TRUE)
}
\arguments{
\item{T}{equitable transform info: output from transformE}

\item{num}{5 default number of indiviual plot to be made}

\item{signal}{2D data set representing the signal. Must be the same size as T$smat}

\item{limits}{default NULL limits on yaxis of plots and range for rainbow colouring in images}

\item{xlimits}{default NULL limits on xaxis of plots}

\item{indiv}{default FALSE: plots individual row or column dependendent on transpose flag}

\item{columns}{default FALSE TRUE:plot all columns on one plot}

\item{images}{default TRUE: plots false colour images}

\item{density}{ignore}

\item{versus}{plots all individual vs reference default FALSE:}

\item{xvsref}{column to be used as reference against which all others are plotted default NULL}

\item{yline}{default 3 lines out from plot to display yaxis values}

\item{yma}{default 5 cahracters out from plot to display ylabel}

\item{transpose}{default FALSE TRUE: plot all rows on one plot}

\item{of}{according to above flags plots original data default FALSE}

\item{lf}{according to above flags plots least squares transformdata default FALSE}

\item{ef}{according to above flags plots equitable transform data default FALSE}

\item{avef}{according to above flags plots equitbale transform data formed using only average profile default FALSE}

\item{errb}{according to above flags uses error bars when possible default FALSE}

\item{row_unit}{row axis label  default Row Number}

\item{col_unit}{column axis label  default Row Number}

\item{z_unit}{label  for quantity measured in data}

\item{genname}{main title to be included}

\item{stderror}{according to above flags uses error bars of standard error rather than standard deviaiton when possible default FALSE}

\item{fcontour}{default TRUE: plots contour maps}
}
\value{
None
}
\description{
Uses functions plotimages plotindiv plotO_S_E_lscol plotversus plot_vsref.
The function plotsummary uses this function to summarize the transform data
}
\examples{
d<-eg4(2,2)
Td<-transformE(d)
d_noise<-d+rnorm(prod(dim(d)),mean=0,sd=(1/4*sd(d,na.rm=TRUE)))
Td_noise<-transformE(d_noise)
plotsome(T=Td_noise,lf=TRUE,of=TRUE)
# default only plots imagef of signal original and equitable
plotsome(T=Td_noise,signal=Td$smat,of=TRUE)
plotsome(T=Td_noise,signal=Td$smat,indiv=TRUE,xvsref=ncol(Td$smat))
plotsome(T=Td_noise,signal=Td$smat,columns=TRUE,images=FALSE,
lf=TRUE,of=TRUE)
plotsome(T=Td_noise,signal=Td$smat,columns=TRUE,images=FALSE,
transpose=TRUE,lf=TRUE,of=TRUE)
plotsome(T=Td_noise,signal=Td$smat,indiv=TRUE,of=TRUE,lf=TRUE)
plotsome(T=Td_noise,signal=Td$smat,indiv=TRUE,of=TRUE,errb=TRUE)
plotsome(T=Td_noise,signal=Td$smat,indiv=TRUE,of=TRUE,errb=TRUE,stderror=TRUE)
#num= NULL  all individuals are plotted
plotsome(T=Td_noise,signal=Td$smat,indiv=TRUE,of=TRUE,num=NULL)
# plots images and transforms vs original and signal
plotsome(T=Td_noise,signal=Td$smat,of=TRUE,lf=TRUE,versus=TRUE)


}
