% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{runstatsNS}
\alias{runstatsNS}
\title{mean and std dev of transform differenced with the signal data set}
\usage{
runstatsNS(Tx, T_noise, extraf = FALSE)
}
\arguments{
\item{Tx}{signal transform information from transformE}

\item{T_noise}{noisy transform information from transformE}

\item{extraf}{ignore}
}
\value{
None
}
\description{
Finds mean and standard deviations of transforms and residuals between transfgorms and signal
}
\examples{
#first construct transfor of data and trnasform of signal
d<-eg4(2,2)
Td<-transformE(d)
d_noise<-d+rnorm(prod(dim(d)),mean=0,sd=(1/4*sd(d,na.rm=TRUE)))
Td_noise<-transformE(d_noise)
runstatsNS(Tx=Td_noise,T=Td)  #

}
