% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equitable_Transform.R
\name{xvsrefplot}
\alias{xvsrefplot}
\title{Plots individual sequences against a reference sequence}
\usage{
xvsrefplot(Td, cgroup = NULL, ylim = NULL, ref = NULL, br = "",
  pf = TRUE, extranames = NULL, err95 = TRUE, fitl = TRUE)
}
\arguments{
\item{Td}{transform information from transformE}

\item{cgroup}{cgroup is the vector of columns to be plotted. Default NULL;10 colums across matrix are plotted}

\item{ylim}{vector of min and max for plots to display default (NULL) function calculates same for all}

\item{ref}{reference column against which all columns are plotted. Default NULL; column index with largest variation}

\item{br}{text for headings of plots}

\item{pf}{plotflag default TRUE}

\item{extranames}{extra names to use for legend of t points}

\item{err95}{2*std. dev/sqrt(N) where std deviation on the slope is found when calculating equitable slopes  Default=TRUE}

\item{fitl}{Default TRUE. Fitted line is displayed}
}
\value{
column index with largest variation amongst columns that were plotted
}
\description{
Equitable transform with errors, original and Equitable transform line parameters can be shown
If number of point in variable t is less than 26 then each t point is colured and labelled in the legend
}
\examples{
#first construct transfor of data and transform of signal
d<-eg4(1,2)
d_noise<-d+rnorm(prod(dim(d)),mean=0,sd=(1/4*sd(d,na.rm=TRUE)))
Td_noise<-transformE(d_noise)
xvsrefplot(Td=Td_noise)
xvsrefplot(Td=Td_noise,ref="Row_Ave")
nc<-seq(1,ncol(Td_noise$smat), by=1)
xvsrefplot(Td=Td_noise,cgroup=nc,ref=ncol(Td_noise$smat),
br=paste0( "Equitable Profiles"))
xvsrefplot(Td=Td_noise,cgroup=c(1,4,9,12),ref=7,
br=paste0( "Equitable Profiles"))
xvsrefplot(Td=Td_noise,cgroup=c(1,4,9,12),ref=7,
br=paste0( "Equitable Profiles with std dev"),fitl=FALSE)
xvsrefplot(Td=Td_noise,cgroup=c(12),ref=7,
br=paste0( "Equitable Profiles"))
xvsrefplot(Td=Td_noise,cgroup=c(12),ref=7,
br=paste0( "Equitable Profiles"),err95=FALSE)

d<-eg5(3,3)
d_noise<-d+rnorm(prod(dim(d)),mean=0,sd=(1/2*sd(d,na.rm=TRUE)))
Td_noise<-transformE(d_noise)
xvsrefplot(Td=Td_noise)

}
